requirements.txt
wheel
twine
requests
numpy
pandas
urllib3
progressbar2
pytest
pytest-mock
python-dotenv
tabulate
